# backend/apispec.yaml
openapi: "3.0.3"
info:
  title: "kJuiceApp API"
  description: "kJuiceApp バックエンドの各種操作を提供する"
  version: "0.1.0"
servers:
  - url: "http://localhost/backend"
    description: "local"
...
paths:
...
  /items:
    get:
      tags:
        - items
      summary: Get item list
      description: 商品リストを取得する
      parameters:
        - in: query
          name: isActive
          schema:
            type: boolean
          description: activeが一致する商品のみ取得する．未指定の場合すべての商品を取得する.
        - in: query
          name: category
          schema:
            $ref: "#/components/schemas/ItemCategory"
          description: 指定されたグループの商品のみ取得する．未指定の場合すべての商品を取得する．
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemResponse"
        "400":
          description: 不正なパラメータ
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - items
      summary: Create new item
      description: 新しい商品を追加する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemPost"
      responses:
        "201":
          description: 追加成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          description: 不正なパラメータ
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: ID重複
  /items/{itemId}:
    patch:
      tags:
        - items
      summary: Update item information by ID
      description: 指定したIDの商品情報を更新する
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
          description: 更新対象の商品ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemPatch"
        required: true
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "404":
          description: 指定されたIDが存在しない
    delete:
      tags:
        - items
      summary: Delete item by ID
      description: 指定されたIDの商品を削除する
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
          description: 削除対象の商品ID
      responses:
        "204":
          description: 削除成功
        "404":
          description: 指定されたIDが存在しない
...
  /stats/items/{itemId}:
    get:
      tags:
        - stats
      summary: Get sales stats on specific item
      description: 指定されたIDの商品について売上金額の合計を取得する．パラメータによって集計期間や集計単位を指定できる.
      parameters:
        - in: path
          name: itemId
          schema:
            type: string
          required: true
        - in: query
          name: start
          schema:
            type: string
            format: date
        - in: query
          name: end
          schema:
            type: string
            format: date
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatisticsOnId"
        "400":
          description: 不正なパラメータ
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 指定されたIDが存在しない
components:
  schemas:
    ItemCategory:
      type: string
      enum:
        - juice
        - food
...
    Error:
      type: object
      properties:
        title:
          type: string
    ItemPost:
      type: object
      properties:
        id:
          type: string
          example: gogotea
        name:
          type: string
          example: 午後の紅茶
        sellingPrice:
          type: integer
          description: 売値
        costPrice:
          type: integer
          description: 仕入れ値
        category:
          type: string
          enum:
            - juice
            - food
      required:
        - id
        - name
        - sellingPrice
        - costPrice
        - category
    ItemPatch:
      type: object
      properties:
        name:
          type: string
          example: 午後の紅茶
        sellingPrice:
          type: integer
          description: 売値
        costPrice:
          type: integer
          description: 仕入れ値
        category:
          type: string
          enum:
            - juice
            - food
        isActive:
          type: boolean
    ItemResponse:
      type: object
      properties:
        id:
          type: string
          example: gogotea
        name:
          type: string
          example: 午後の紅茶
        sellingPrice:
          type: integer
          description: 売値
        costPrice:
          type: integer
          description: 仕入れ値
        category:
          type: string
          enum:
            - juice
            - food
        isActive:
          type: boolean
        salesCount:
          type: integer
...
    StatisticElement:
      type: object
      properties:
        label:
          type: string
        count:
          type: integer
          description: 購入回数
        amount:
          type: integer
          description: 利用金額
    StatisticsOnId:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        totalCount:
          type: integer
        totalAmount:
          type: integer
        elements:
          type: array
          items:
            $ref: "#/components/schemas/StatisticElement"